src/requirements.txt
cython==0.29.1
pillow==5.3.0
pytesseract==0.2.4
tesserocr==0.2.5